Polish up based on feedback
