+++ rest/security/src/main/java/example/company/SecurityConfiguration.java
@@ -13,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package example.company;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,10 +24,10 @@ import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 
 /**
- * This application is secured at both the URL level for some parts, and the method level
- * for other parts. The URL security is shown inside this code, while method-level annotations
- * are enabled at by
- * {@link org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity EnableGlobalMethodSecurity}
+ * This application is secured at both the URL level for some parts, and the method level for other parts. The URL
+ * security is shown inside this code, while method-level annotations are enabled at by
+ * {@link org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity
+ * EnableGlobalMethodSecurity}
  *
  * @author Greg Turnquist
  */
@@ -37,8 +36,7 @@ import org.springframework.security.config.annotation.web.configuration.WebSecur
 public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
 
 	/**
-	 * This section defines the user accounts which can be used for authentication as well as
-	 * the roles each user has.
+	 * This section defines the user accounts which can be used for authentication as well as the roles each user has.
 	 *
 	 * @param auth
 	 * @throws Exception
@@ -51,12 +49,9 @@ public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
 	}
 
 	/**
-	 * This section defines the security policy for the app.
-	 * - BASIC authentication is supported (enough for this REST-based demo)
-	 * - /employees is secured using URL security shown below
-	 * - CSRF headers are disabled since we are only testing the REST interface, not a web one.
-	 *
-	 * NOTE: GET is not shown which defaults to permitted.
+	 * This section defines the security policy for the app. - BASIC authentication is supported (enough for this
+	 * REST-based demo) - /employees is secured using URL security shown below - CSRF headers are disabled since we are
+	 * only testing the REST interface, not a web one. NOTE: GET is not shown which defaults to permitted.
 	 *
 	 * @param http
 	 * @throws Exception
