--- rest/security/src/test/resources/logback.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration>
-	<include resource="org/springframework/boot/logging/logback/base.xml"/>
-	<!--<logger name="org.springframework.web" level="DEBUG"/>-->
-</configuration>
